Move metadata.ts to components folder
