+++ app/not-found.tsx
@@ -1,5 +1,5 @@
+import getMetadata from '@/components/metadata';
 import Link from 'next/link';
-import getMetadata from './res/metadata';
 
 const title = '404: This didn\'t work.';
